+++ src/types-internal/editor-modules.d.ts
@@ -9,7 +9,6 @@ import Notifier from '../components/modules/notifier';
 import DragNDrop from '../components/modules/dragNDrop';
 import ModificationsObserver from '../components/modules/modificationsObserver';
 import Renderer from '../components/modules/renderer';
-import Sanitizer from '../components/modules/sanitizer';
 import Tools from '../components/modules/tools';
 import API from '../components/modules/api/index';
 import Caret from '../components/modules/caret';
@@ -49,7 +48,6 @@ export interface EditorModules {
   DragNDrop: DragNDrop;
   ModificationsObserver: ModificationsObserver;
   Renderer: Renderer;
-  Sanitizer: Sanitizer;
   Tools: Tools;
   API: API;
   Caret: Caret;