+++ src/components/block/index.ts
@@ -46,11 +46,6 @@ interface BlockConstructorOptions {
   readOnly: boolean;
 
   /**
-   * Tunes for current Block
-   */
-  tunes: ToolsCollection<BlockTune>;
-
-  /**
    * Tunes data for current Block
    */
   tunesData: {[name: string]: BlockTuneData};
@@ -224,7 +219,6 @@ export default class Block {
     tool,
     api,
     readOnly,
-    tunes,
     tunesData,
   }: BlockConstructorOptions) {
     this.name = tool.name;
@@ -241,7 +235,7 @@ export default class Block {
     /**
      * @type {BlockTune[]}
      */
-    this.tunes = tunes;
+    this.tunes = tool.tunes;
 
     this.composeTunes(tunesData);