+++ src/components/modules/api/sanitizer.ts
@@ -1,6 +1,7 @@
-import { Sanitizer } from '../../../../types/api';
+import { Sanitizer as ISanitizer } from '../../../../types/api';
 import { SanitizerConfig } from '../../../../types/configs';
 import Module from '../../__module';
+import { clean } from '../../utils/sanitizer';
 
 /**
  * @class SanitizerAPI
@@ -12,7 +13,7 @@ export default class SanitizerAPI extends Module {
    *
    * @returns {Sanitizer}
    */
-  public get methods(): Sanitizer {
+  public get methods(): ISanitizer {
     return {
       clean: (taintString, config): string => this.clean(taintString, config),
     };
@@ -27,6 +28,6 @@ export default class SanitizerAPI extends Module {
    * @returns {string}
    */
   public clean(taintString: string, config: SanitizerConfig): string {
-    return this.Editor.Sanitizer.clean(taintString, config);
+    return clean(taintString, config);
   }
 }