+++ test/cypress/support/index.d.ts
@@ -25,7 +25,20 @@ declare global {
        */
       paste(data: {[type: string]: string}): Chainable<Subject>
 
+      /**
+       * Copy command to dispatch copy event on subject
+       *
+       * @usage
+       * cy.get('div').copy().then(data => {})
+       */
       copy(): Chainable<{ [type: string]: any }>;
+
+      /**
+       * Cut command to dispatch cut event on subject
+       *
+       * @usage
+       * cy.get('div').cut().then(data => {})
+       */
       cut(): Chainable<{ [type: string]: any }>;
     }