+++ src/components/modules/caret.ts
@@ -531,7 +531,7 @@ export default class Caret extends Module {
      * If there is no child node, append empty one
      */
     if (fragment.childNodes.length === 0) {
-      fragment.appendChild(new Text(''));
+      fragment.appendChild(new Text());
     }
 
     const lastChild = fragment.lastChild;