+++ docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### 2.20.2
+
+- `Fix` - Use `<div>` tag instead of `<p>` in the copied HTML data [#1645](https://github.com/codex-team/editor.js/issues/1645).
+
 ### 2.20.1
 
 - `Fix` - Create a new block when clicked at the bottom [#1588](https://github.com/codex-team/editor.js/issues/1588).
@@ -13,7 +17,6 @@
 ### 2.20.0
 
 - `New` — [Block Tunes API](block-tunes.md) added
-
 
 ### 2.19.3